feat : finishing repo and handler logic

+++ model/campaign/campaign_model.go
@@ -15,13 +15,13 @@ type Campaign struct {
 
 // Campaign related services
 type CampaignService interface {
-	GetByID(ctx context.Context, ID int) (*Campaign, error)
-	GetByListID(ctx context.Context, ListID []int) (*[]Campaign, error)
+	GetByID(ctx context.Context, ID int64) (*Campaign, error)
+	GetByListID(ctx context.Context, ListID []int64) (*[]Campaign, error)
 }
 
 // Campaign related repository
 
 type CampaignRepository interface {
-	GetByID(ctx context.Context, ID int) (*Campaign, error)
-	GetByListID(ctx context.Context, ListID []int) (*[]Campaign, error)
+	GetByID(ctx context.Context, ID int64) (*Campaign, error)
+	GetByListID(ctx context.Context, ListID []int64) (*[]Campaign, error)
 }
